finished refactoring services following.js
-finished changing over to different request building method for following services which were causing bugs on firefox
