+++ epoch_frontend/src/services/following.js
@@ -9,20 +9,44 @@ function getAccountList() {
     if (!session_id) {
         return;
     }
-    const currentLocation = window.location;
-    const serverUrl = `${currentLocation.protocol}//${currentLocation.hostname}:8080`;
-    const url = `${serverUrl}/api/follow/accountList/`;
-    const headers = {credentials: 'include'}
+    return new Promise((resolve, reject) => {
+        var http = new XMLHttpRequest();
+        const currentLocation = window.location;
+        const serverUrl = `${currentLocation.protocol}//${currentLocation.hostname}:8080`;
+        http.open("GET",`${serverUrl}/api/follow/accountList/`, true );
+        http.setRequestHeader("Content-Type", "application/json");
+        http.withCredentials = true;
+        http.timeout = 10000;
 
-    const data = fetch(url, headers)
-        .then(response => response.json())
-        .catch(error => {
-            //this.setState({ errorMessage: error.toString() }); // this was causing an error in rendering the rest of the profile so i commented it for now, console.error should be enough - Ayman
-            console.error('There was an error getting users lists!', error);
-            return false;
-        });
+        http.onreadystatechange = function () {
+            if (http.readyState === 4) {
+                if (http.status === 200) {
+                    const userList = JSON.parse(http.responseText);
+                    resolve(userList);
+                } else {
+                    if (http.status !== 0) {
+                        reject(http.statusText);
+                    } else {
+                        reject("Connection refused: The server is not running or unreachable");
+                    }
+                }
+            }
+        };
 
-    return(data)
+        http.ontimeout = function () {
+            reject("Request timed out");
+        }
+
+        http.onerror = function () {
+            reject("An error occurred sending account list request");
+        }
+
+        http.onabort = function () {
+            reject("Account List Request aborted");
+        }
+
+        http.send(JSON.stringify({session_id: session_id}));
+    });
 }
 
 function getFollowingList() {
@@ -30,72 +54,157 @@ function getFollowingList() {
     if (!session_id) {
         return;
     }
-    const currentLocation = window.location;
-    const serverUrl = `${currentLocation.protocol}//${currentLocation.hostname}:8080`;
-    const url = `${serverUrl}/api/follow/followingList/`;
-    const headers = {credentials: 'include'}
+    return new Promise((resolve, reject) => {
+        var http = new XMLHttpRequest();
+        const currentLocation = window.location;
+        const serverUrl = `${currentLocation.protocol}//${currentLocation.hostname}:8080`;
+        http.open("GET",`${serverUrl}/api/follow/followingList/`, true );
+        http.setRequestHeader("Content-Type", "application/json");
+        http.withCredentials = true;
+        http.timeout = 10000;
 
-    const data = fetch(url, headers)
-        .then(response => response.json())
-        .catch(error => {
-            //this.setState({ errorMessage: error.toString() }); // this was causing an error in rendering the rest of the profile so i commented it for now, console.error should be enough - Ayman
-            console.error('There was an error getting following lists!', error);
-            return false;
-        });
+        http.onreadystatechange = function () {
+            if (http.readyState === 4) {
+                if (http.status === 200) {
+                    const followingList = JSON.parse(http.responseText);
+                    resolve(followingList);
+                } else {
+                    if (http.status !== 0) {
+                        reject(http.statusText);
+                    } else {
+                        reject("Connection refused: The server is not running or unreachable");
+                    }
+                }
+            }
+        };
 
-    return(data)    
+        http.ontimeout = function () {
+            reject("Request timed out");
+        }
+
+        http.onerror = function () {
+            reject("An error occurred sending following list request");
+        }
+
+        http.onabort = function () {
+            reject("following List Request aborted");
+        }
+
+        http.send(JSON.stringify({session_id: session_id}));
+    });
+
 }
 
-function followAccount(userToFollow) {
+async function fillUserList() {
+    var users = await getAccountList();
+    var following = await getFollowingList();
+
+    for (var i in users) {
+        users[i].isFollowing = false;
+    }
+    for (var k in following) {
+        for (var j in users) {
+            if (users[j].user_id === following[k].following_id){
+                users[j].isFollowing = true;
+            }
+        }
+    }
+    if(users.length>0){
+        users.sort(function(a,b){
+            return b.isFollowing - a.isFollowing;
+        });
+    }
+    return users;
+}
+
+function followAccount(target) {
     const session_id = getCookie('epoch_session_id');
     if (!session_id) {
         return;
     }
-    var params = {
-        session_id: session_id,
-        userToFollow: userToFollow
-    }
-    const currentLocation = window.location;
-    const serverUrl = `${currentLocation.protocol}//${currentLocation.hostname}:8080`;
-    const url = `${serverUrl}/api/follow/follow/`;
+    return new Promise((resolve, reject) => {
+        var http = new XMLHttpRequest();
+        const currentLocation = window.location;
+        const serverUrl = `${currentLocation.protocol}//${currentLocation.hostname}:8080`;
+        http.open("POST",`${serverUrl}/api/follow/follow/`, true );
+        http.setRequestHeader("Content-Type", "application/json");
+        http.withCredentials = true;
+        http.timeout = 10000;
 
-    fetch(url,{
-        method: 'POST',
-        mode: 'cors',
-        credentials:'include',
-        body: JSON.stringify(params)
-    })
-        .catch(error => {
-            //this.setState({ errorMessage: error.toString() }); // this was causing an error in rendering the rest of the profile so i commented it for now, console.error should be enough - Ayman
-            console.error('There was an error following an account!', error);
-            return false;
-        });
+        http.onreadystatechange = function () {
+            if (http.readyState === 4) {
+                if (http.status === 200) {
+                    resolve(true);
+                } else {
+                    if (http.status !== 0) {
+                        reject(http.statusText);
+                    } else {
+                        reject("Connection refused: The server is not running or unreachable");
+                    }
+                }
+            }
+        };
+
+        http.ontimeout = function () {
+            reject("Request timed out");
+        }
+
+        http.onerror = function () {
+            reject("An error occurred sending following list request");
+        }
+
+        http.onabort = function () {
+            reject("following List Request aborted");
+        }
+
+        http.send(JSON.stringify({session_id: session_id, userToFollow: target}));
+    });
+
 }
 
-function unfollowAccount(userToUnfollow) {
+function unfollowAccount(target) {
     const session_id = getCookie('epoch_session_id');
     if (!session_id) {
         return;
     }
-    var params = {
-        session_id: session_id,
-        userToUnfollow: userToUnfollow
-    }
-    const currentLocation = window.location;
-    const serverUrl = `${currentLocation.protocol}//${currentLocation.hostname}:8080`;
-    const url = `${serverUrl}/api/follow/unfollow/`;
-    fetch(url,{
-        method: 'POST',
-        mode: 'cors',
-        credentials:'include',
-        body: JSON.stringify(params)
-    })
-        .catch(error => {
-            //this.setState({ errorMessage: error.toString() }); // this was causing an error in rendering the rest of the profile so i commented it for now, console.error should be enough - Ayman
-            console.error('There was an error unfollowing an account!', error);
-            return false;
-        });
+    return new Promise((resolve, reject) => {
+        var http = new XMLHttpRequest();
+        const currentLocation = window.location;
+        const serverUrl = `${currentLocation.protocol}//${currentLocation.hostname}:8080`;
+        http.open("POST",`${serverUrl}/api/follow/unfollow/`, true );
+        http.setRequestHeader("Content-Type", "application/json");
+        http.withCredentials = true;
+        http.timeout = 10000;
+
+        http.onreadystatechange = function () {
+            if (http.readyState === 4) {
+                if (http.status === 200) {
+                    resolve(true);
+                } else {
+                    if (http.status !== 0) {
+                        reject(http.statusText);
+                    } else {
+                        reject("Connection refused: The server is not running or unreachable");
+                    }
+                }
+            }
+        };
+
+        http.ontimeout = function () {
+            reject("Request timed out");
+        }
+
+        http.onerror = function () {
+            reject("An error occurred sending following list request");
+        }
+
+        http.onabort = function () {
+            reject("following List Request aborted");
+        }
+
+        http.send(JSON.stringify({session_id: session_id, userToUnfollow: target}));
+    });
 
 }
 
-module.exports = {getAccountList, getFollowingList, followAccount, unfollowAccount}
+module.exports = {followAccount, unfollowAccount, fillUserList, getFollowingList}
